Fixed WorkShop compilation warnings
git-svn-id: https://anonsvn.ncbi.nlm.nih.gov/repos/v1/trunk/c++@32787 78c7ea69-d796-4a43-9a09-de51944f1b03

+++ src/corelib/ncbi_system.cpp
@@ -497,9 +497,9 @@ bool GetMemoryUsage(size_t* total, size_t* resident, size_t* shared)
 #elif defined(NCBI_OS_SOLARIS)
     Int8 len = CFile("/proc/self/as").GetLength();
     if (len > 0) {
-        *total    = len;
-        *resident = len; // conservative estimate
-        *shared   = 0;   // does this info exist anywhere?
+        *total    = (size_t)len;
+        *resident = (size_t)len; // conservative estimate
+        *shared   = 0;           // does this info exist anywhere?
         return true;
     }
 #elif defined(HAVE_GETRUSAGE)
